EmporiumMall1/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.0.0
django-dotenv==1.4.2
fontawesomefree==6.4.0
mysqlclient==2.2.0
Pillow==8.2.0
sqlparse==0.4.4
typing_extensions==4.7.1
